ADD: edit actor and confirm actor deletion

+++ app/assets/javascripts/controllers/actor_controller.js
@@ -1,13 +1,21 @@
 App.ActorController = Ember.ObjectController.extend({
   isShowingDetails: false,
+  isEditing: false,
   toggleDetails: function() {
     this.set('isShowingDetails', !this.get('isShowingDetails'));
   },
-  saveNew: function() {
+  toggleEdit: function() {
+    this.set("isEditing", !this.get('isEditing'));
+  },
+  save: function() {
+    this.set("isEditing", false);
+    this.set("isShowingDetails", false);
     this.get('store').commit();
   },
   delete: function() {
-    this.get('model').deleteRecord();
-    this.get('store').commit();
+    if (confirm("Are you sure to dele the actor?")) {
+      this.get('model').deleteRecord();
+      this.get('store').commit();
+    };
   }
 });
